% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnns.R
\name{bnns_params}
\alias{bnns_params}
\alias{L}
\alias{warmup}
\alias{chains}
\alias{iter}
\alias{nodes}
\alias{act_fn}
\title{Parameter functions for Bayesian Neural Networks}
\usage{
L(range = c(1L, 5L), trans = NULL)

warmup(range = c(100L, 1000L), trans = NULL)

chains(range = c(1L, 4L), trans = NULL)

iter(range = c(500L, 2000L), trans = NULL)

nodes(range = c(1L, 64L), trans = NULL)

act_fn(values = c("tanh", "sigmoid", "softplus", "relu", "linear"))
}
\arguments{
\item{range}{A two-element vector holding the defaults for the smallest and largest possible values.}

\item{trans}{A \code{trans} object from the \code{scales} package, could be \code{NULL}.}

\item{values}{A character vector of possible values.}
}
\value{
A \code{quant_param} or \code{qual_param} object from the \code{dials} package.
}
\description{
These functions provide \code{dials} parameter objects for tuning
the hyperparameters of Bayesian Neural Networks.
}
