% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnns.R
\name{bnns_train}
\alias{bnns_train}
\title{Internal function for training the BNN}
\usage{
bnns_train(
  train_x,
  train_y,
  L = 1,
  nodes = 16,
  act_fn = 2,
  out_act_fn = 1,
  iter = 1000,
  warmup = 200,
  thin = 1,
  chains = 2,
  cores = 2,
  seed = 123,
  ...
)
}
\arguments{
\item{train_x}{A numeric matrix representing the input features (predictors) for training. Rows correspond to observations, and columns correspond to features.}

\item{train_y}{A numeric vector representing the target values for training. Its length must match the number of rows in \code{train_x}.}

\item{L}{An integer specifying the number of hidden layers in the neural network. Default is 1.}

\item{nodes}{An integer or vector specifying the number of nodes in each hidden layer. If a single value is provided, it is applied to all layers. Default is 16.}

\item{act_fn}{An integer or vector specifying the activation function(s) for the hidden layers. Options are:
\itemize{
\item \code{1} for tanh
\item \code{2} for sigmoid (default)
\item \code{3} for ReLU
\item \code{4} for softplus
}}

\item{out_act_fn}{An integer specifying the activation function for the output layer. Options are:
\itemize{
\item \code{1} for linear (default)
\item \code{2} for sigmoid
\item \code{3} for softmax
}}

\item{iter}{An integer specifying the total number of iterations for the Stan sampler. Default is \code{1e3}.}

\item{warmup}{An integer specifying the number of warmup iterations for the Stan sampler. Default is \code{2e2}.}

\item{thin}{An integer specifying the thinning interval for Stan samples. Default is 1.}

\item{chains}{An integer specifying the number of Markov chains. Default is 2.}

\item{cores}{An integer specifying the number of CPU cores to use for parallel sampling. Default is 2.}

\item{seed}{An integer specifying the random seed for reproducibility. Default is 123.}

\item{...}{Currently not in use.}
}
\value{
An object of class \code{"bnns"} containing the following components:
\describe{
\item{\code{fit}}{The fitted Stan model object.}
\item{\code{call}}{The matched call.}
\item{\code{data}}{A list containing the Stan data used in the model.}
}
}
\description{
This function performs the actual fitting of the Bayesian Neural Network.
It is called by the exported bnns methods and is not intended for direct use.
}
\details{
The function uses the \code{generate_stan_code} function to dynamically generate Stan code based on the specified number of layers and nodes. Stan is then used to fit the Bayesian Neural Network.
}
\examples{
# Example usage:
train_x <- matrix(runif(20), nrow = 10, ncol = 2)
train_y <- rnorm(10)
model <- bnns:::bnns_train(train_x, train_y, L = 1, nodes = 2, act_fn = 2,
iter = 1e2, warmup = 5e1, chains = 1)

# Access Stan model fit
model$fit

}
\seealso{
\code{\link[rstan]{stan}}
}
\keyword{internal}
