% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stan_code.R
\name{generate_stan_code}
\alias{generate_stan_code}
\title{Generate the required Stan code}
\usage{
generate_stan_code(num_layers, nodes)
}
\arguments{
\item{num_layers}{Number of hidden layers.}

\item{nodes}{A vector of number of nodes in each hidden layer.}
}
\value{
A string with the required Stan code for Bayesian neural network.
}
\description{
Generate the required Stan code
}
\examples{
generate_stan_code(3, c(16, 4, 2))
generate_stan_code(2, c(4, 2))
}
