% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_bnns}
\alias{load_bnns}
\title{Load a fitted bnns model from disk}
\usage{
load_bnns(file)
}
\arguments{
\item{file}{A character string specifying the path to the saved model.}
}
\value{
A fitted \code{bnns} object.
}
\description{
Load a fitted bnns model from disk
}
