% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnns.R
\name{loo.bnns}
\alias{loo.bnns}
\title{Leave-One-Out Cross-Validation (LOO) for bnns models}
\usage{
\method{loo}{bnns}(x, ...)
}
\arguments{
\item{x}{A fitted \code{bnns} model object.}

\item{...}{Additional arguments passed to \code{loo::loo()}.}
}
\value{
A \code{loo} object containing model comparison metrics.
}
\description{
Leave-One-Out Cross-Validation (LOO) for bnns models
}
