% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_functions.R
\name{relu}
\alias{relu}
\title{relu transformation}
\usage{
relu(x)
}
\arguments{
\item{x}{A numeric vector or matrix on which relu transformation is going to be applied.........................}
}
\value{
A numeric vector or matrix after relu transformation.
}
\description{
relu transformation
}
\examples{
\dontrun{
relu(matrix(1:4), nrow = 2)
}
}
