% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_functions.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid transformation}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{A numeric vector or matrix on which sigmoid transformation is going to be applied.........................}
}
\value{
A numeric vector or matrix after sigmoid transformation.
}
\description{
sigmoid transformation
}
\examples{
\dontrun{
sigmoid(matrix(1:4), nrow = 2)
}
}
