% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_functions.R
\name{softplus}
\alias{softplus}
\title{softplus transformation}
\usage{
softplus(x)
}
\arguments{
\item{x}{A numeric vector or matrix on which softplus transformation is going to be applied.........................}
}
\value{
A numeric vector or matrix after softplus transformation.
}
\description{
softplus transformation
}
\examples{
\dontrun{
softplus(matrix(1:4), nrow = 2)
}
}
