% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation_functions.R
\name{tanh}
\alias{tanh}
\title{tanh transformation}
\usage{
tanh(x)
}
\arguments{
\item{x}{A numeric vector or matrix on which tanh transformation is going to be applied.........................}
}
\value{
A numeric vector or matrix after tanh transformation.
}
\description{
tanh transformation
}
\examples{
\dontrun{
tanh(matrix(1:4), nrow = 2)
}
}
