% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnns_parsnip_helpers.R
\name{translate_activation}
\alias{translate_activation}
\title{Translate parsnip activation name to bnns code}
\usage{
translate_activation(activation)
}
\description{
Translate parsnip activation name to bnns code
}
\keyword{internal}
