% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnns_parsnip_helpers.R
\name{translate_out_activation}
\alias{translate_out_activation}
\title{Translate output activation name to bnns code}
\usage{
translate_out_activation(out_act_fn)
}
\description{
Translate output activation name to bnns code
}
\keyword{internal}
